+++ paper/papercode/AverageTimingPlot.R
@@ -57,7 +57,7 @@
 
  par(fig=c(0,1,0,1),oma=c(0,0,0,0),mar=c(0,0,0,0),new=TRUE)
  plot(0,0,type="n",bty="n",xaxt="n",yaxt="n")
- legend("bottom",c("QuanBiMo","LPAwb+","Exhaustive LPAwb+"),pch=c(1,3,4),col=c("black","grey30","grey65"),bty="n",horiz=TRUE)
+ legend("bottom",c("QuanBiMo","LPAwb+","DIRTLPAwb+"),pch=c(1,3,4),col=c("black","grey30","grey65"),bty="n",horiz=TRUE)
 
  dev.off()